+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Hex color support for `color` and `bgcolor`
 - Test system modified
 - Random mode modified
 ## [0.6] - 2022-04-13